+++ ITStudy/Models/Vouchers.cs
@@ -7,7 +7,7 @@ public class Vouchers
     public Student? Student { get; set; }
     public Courses? Course { get; set; }
     public string Title { get; set; } = string.Empty;
-    public long Discount { get; set; } = 0;
+    public double Discount { get; set; } = 0;
     public long? CourseId { get; set; } = null;
     public DateTime? Public_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
     public DateTime? Expire_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
@@ -21,7 +21,7 @@ public enum Status_Vouchers
 public class Vouchers_Add
 {
     public string Title { get; set; } = string.Empty;
-    public int Discount { get; set; } = 0;
+    public double Discount { get; set; } = 0;
     public DateTime? Public_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
     public DateTime? Expire_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
     public Status_Vouchers Status { get; set; } = Status_Vouchers.active;
@@ -31,7 +31,7 @@ public class Vouchers_Get
 {
     public long Id { get; set; }
     public string Title { get; set; } = string.Empty;
-    public int Discount { get; set; } = 0;
+    public double Discount { get; set; } = 0;
     public long? CourseId { get; set; } = null;
     public DateTime? Public_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
     public DateTime? Expire_Date { get; set; } = DateTime.UtcNow.ToUniversalTime();
